added service layer, app layer and started app for the first time. milestone. Only provider has tests at this point

+++ 06-githubapi/src/api/domain/github/github-error.go
@@ -8,7 +8,7 @@ type GithubError struct {
 }
 
 type GithubErrorResponse struct {
-	StatusCode       int64  `json:"status_code"`
+	StatusCode       int    `json:"status_code"`
 	Message          string `json:"message"`
 	DocumentationUrl string `json:"documentation_url"`
 	Errors           []GithubError
